Fix broken error notification spec for Electron 3.0
This change modifies the use of `fs.readFile` to generate an unhandled error
that does that does not get displayed by the notifications package.  Between
Electron 2.0 and 3.0, Node.js' `fs.readFile` function has now been updated to
throw an exception when a callback isn't supplied so we now have to go a little
further to generate the desired unhandled exception that validates the expected
behavior.